LayoutTests/accessibility/mac/async-increment-decrement-action.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body id="body">

<input type="range" min="0" max="100" value="25" step="25" id="range1"/>

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that increment and decrement actions are async.");

    if (window.accessibilityController) {
        jsTestIsAsync = true;

        var obj = accessibilityController.accessibleElementById("range1");
        
        shouldBe("obj.intValue", "25");

        obj.asyncIncrement();
        shouldBe("obj.intValue", "25");

        setTimeout(function() {
            shouldBe("obj.intValue", "50");

            obj.asyncDecrement();
            shouldBe("obj.intValue", "50");

            setTimeout(function() {
                shouldBe("obj.intValue", "25");
                finishJSTest();
            }, 5);
        }, 5);
    }

</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
